fix from Ali's feedback : clean test code and move transaction start to home
