+++ src/features/home/pages/Home.native.test.tsx
@@ -19,17 +19,12 @@ const mockUseHomepageData = useHomepageData as jest.Mock
 
 jest.mock('libs/geolocation')
 
-const mockStartTransaction = jest.fn((s: string) => {
-  s.toUpperCase()
-})
-const mockFinishTransaction = jest.fn((s: string) => {
-  s.toUpperCase()
-})
+const mockStartTransaction = jest.fn()
+const mockFinishTransaction = jest.fn()
 jest.mock('shared/performance/transactions', () => {
   const originalModule = jest.requireActual('shared/performance/transactions')
 
   return {
-    __esModule: true,
     ...originalModule,
     startTransaction: (s: string) => {
       mockStartTransaction(s)
